requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib[bcrypt]
python-jose
